fix: move wit_* packages into wit-bindgen dir
