+++ docs/wit-bindgen/wit_async.html
@@ -1,0 +1,10 @@
+<html>
+	<head>
+		<title>go.bytecodealliance.org/wit-bindgen/wit_async</title>
+		<meta name="go-import" content="go.bytecodealliance.org/wit-bindgen git https://github.com/bytecodealliance/componentize-go" />
+		<meta http-equiv="refresh" content="0; url=https://pkg.go.dev/go.bytecodealliance.org/wit-bindgen/wit_async">
+	</head>
+	<body>
+		<a href="https://pkg.go.dev/go.bytecodealliance.org/wit-bindgen/wit_async">Redirecting to documentation…</a>
+	</body>
+</html>
