--- docs/wit_async.html
@@ -1,10 +1,0 @@
-<html>
-	<head>
-		<title>go.bytecodealliance.org/wit_async</title>
-		<meta name="go-import" content="go.bytecodealliance.org/wit_async git https://github.com/bytecodealliance/componentize-go" />
-		<meta http-equiv="refresh" content="0; url=https://pkg.go.dev/go.bytecodealliance.org/wit_async">
-	</head>
-	<body>
-		<a href="https://pkg.go.dev/go.bytecodealliance.org/wit_async">Redirecting to documentation…</a>
-	</body>
-</html>
