--- docs/wit_runtime.html
@@ -1,10 +1,0 @@
-<html>
-	<head>
-		<title>go.bytecodealliance.org/wit_runtime</title>
-		<meta name="go-import" content="go.bytecodealliance.org/wit_runtime git https://github.com/bytecodealliance/componentize-go" />
-		<meta http-equiv="refresh" content="0; url=https://pkg.go.dev/go.bytecodealliance.org/wit_runtime">
-	</head>
-	<body>
-		<a href="https://pkg.go.dev/go.bytecodealliance.org/wit_runtime">Redirecting to documentation…</a>
-	</body>
-</html>
