--- docs/wit_types.html
@@ -1,10 +1,0 @@
-<html>
-	<head>
-		<title>go.bytecodealliance.org/wit_types</title>
-		<meta name="go-import" content="go.bytecodealliance.org/wit_types git https://github.com/bytecodealliance/componentize-go" />
-		<meta http-equiv="refresh" content="0; url=https://pkg.go.dev/go.bytecodealliance.org/wit_types">
-	</head>
-	<body>
-		<a href="https://pkg.go.dev/go.bytecodealliance.org/wit_types">Redirecting to documentation…</a>
-	</body>
-</html>
