+++ wit-bindgen/wit_types/wit_option.go
