+++ wit-bindgen/wit_types/wit_result.go
